AllFine

+++ LuoGu/P2704.cpp
@@ -4,23 +4,23 @@
 
 using namespace std;
 
-int n, m, ans, f[2][1048576];
+LL n, m, ans, f[2][1048576];
 bool b, a[101][11];
 char ch;
 
 int main() {
   cin >> n >> m;
-  for (int i = 0; i < n; i++) {
-    for (int j = 0; j < m; j++) {
+  for (LL i = 0; i < n; i++) {
+    for (LL j = 0; j < m; j++) {
       cin >> ch;
       a[i][j] = ch == 'P';
     }
   }
-  for (int i = 0; i < n; i++) {
-    for (int j = 0; j < m; j++) {
+  for (LL i = 0; i < n; i++) {
+    for (LL j = 0; j < m; j++) {
       b ^= 1;
       fill(&f[b][0], &f[b][1048576], 0);
-      for (int k = 0; k < (1 << 2 * m); k++) {
+      for (LL k = 0; k < (1 << 2 * m); k++) {
         bool u1 = (1 << m) & k, u2 = 1 & k, l1 = (1 << 2 * m - 2) & k, l2 = (1 << 2 * m - 3) & k;
         f[b][k >> 1] = max(f[b][k >> 1], f[b ^ 1][k]);
         if (!u1 && !u2 && !l1 && !l2 && a[i][j]) {
@@ -29,7 +29,7 @@ int main() {
       }
     }
   }
-  for (int i = 0; i < (1 << m * 2); i++) {
+  for (LL i = 0; i < (1 << m * 2); i++) {
     ans = max(ans, f[b][i]);
   }
   cout << ans << endl;
